HeatMap/requirements.txt
amqp==2.6.1
asgiref==3.2.7
billiard==3.6.3.0
bleach==3.2.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cmake==3.18.2.post1
colorama==0.4.3
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
Django==2.2.7
django-crispy-forms==1.7.2
django-secretballot==1.0.0
django-widget-tweaks==1.4.8
dlib==19.21.0
docutils==0.16
idna==2.10
image==1.5.32
importlib-metadata==1.7.0
imutils==0.5.3
keyring==21.4.0
kiwisolver==1.2.0
kombu==4.6.11
matplot==0.1.9
matplotlib==3.3.2
mysqlclient==1.4.0
networkx==2.4
numpy==1.19.0
opencv-python==4.2.0.34
packaging==20.4
pathlib==1.0.1
Pillow==7.2.0
pkginfo==1.5.0.1
pycparser==2.20
Pygments==2.7.1
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scipy==1.5.2
Shapely==1.7.0
SimpleWebSocketServer==0.1.1
six==1.15.0
sqlparse==0.3.1
tqdm==4.49.0
twine==3.2.0
typing==3.7.4.3
urllib3==1.25.10
ushlex==0.99.1
vine==1.3.0
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.57.0
zipp==3.1.0
